requirements.txt
aiohttp==3.11.13
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytz==2025.1
websockets==15.0.1
fastapi==0.115.8
uvicorn[standard]==0.34.0
aiosqlite==0.21.0
